refactor: 캐러샐 전체 구조 수정
