+++ src/components/home/Home.tsx
@@ -1,6 +1,6 @@
 import type { ViewType } from "@/routes/routes";
 import { MapBanner } from "@/components/home/parts/MapBanner";
-import { TrendingCarouselSection } from "@/components/home/parts/TrendingCarouselSection";
+import { TrendingCarouselSection } from "@/components/home/parts/trending/TrendingCarouselSection";
 import { EndingSoonSection } from "@/components/home/parts/EndingSoonSection";
 import { SeongsuSection } from "@/components/home/parts/SeongsuSection";
 import { CategorySection } from "@/components/home/parts/CategorySection";
