+++ src/components/home/parts/constants/trendingStyle.ts
@@ -1,0 +1,11 @@
+export const APPBAR_HEIGHT = 60;
+export const DRAG_THRESHOLD_PX = 10;
+
+export const SIDE_PADDING = 20;
+export const GAP = 12;
+
+export const CARD_HEIGHT = 430;
+export const CARD_MAX_WIDTH = 340;
+
+export const INACTIVE_SCALE = 0.92;
+export const INACTIVE_OPACITY = 0.55;
